GP-950 (#362)
* GP-950

* GP-950 refactor admin structure
