+++ openedx/features/genplus_features/genplus/admin/__init__.py
@@ -1,0 +1,2 @@
+from . import admin
+from . import school
